backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jai Smith <jksmithnyc@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.10"
pandas = ">=1.3.5,<2.0.0"
numpy = ">=1.21.6,<1.24.0"
requests = "^2.31.0"
pytz = "^2023.3"
python-dotenv = "^1.0.0"
scipy = "^1.11.1"
neuralprophet = "^1.0.0rc3"
ish-parser = "^0.0.25"
psycopg2-binary = "^2.9.7"
boto3 = "^1.28.21"
botocore = "^1.28.21"
s3fs = "^0.4"
torch = [
  { url = "https://download.pytorch.org/whl/cpu/torch-1.13.1%2Bcpu-cp310-cp310-linux_x86_64.whl", markers = "sys_platform == 'linux'" },
  { url = "https://download.pytorch.org/whl/cpu/torch-1.13.1-cp310-none-macosx_11_0_arm64.whl", markers = "sys_platform == 'darwin'" }
]
awslambdaric = "^2.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
